+++ package.json
@@ -85,12 +85,12 @@
     "@commitlint/cli": "^17.4.4",
     "@commitlint/config-conventional": "^17.4.4",
     "@commitlint/prompt": "^17.4.4",
-    "@lidofinance/browser-service": "1.4.1",
+    "@lidofinance/browser-service": "1.4.3",
     "@lidofinance/discord-reporter": "1.5.2",
     "@lidofinance/eslint-config": "^0.43.0",
     "@lidofinance/lido-ethereum-sdk": "^4.4.0",
     "@lidofinance/wallets-testing-extensions": "1.3.0",
-    "@lidofinance/wallets-testing-wallets": "1.36.1",
+    "@lidofinance/wallets-testing-wallets": "1.36.3",
     "@nestjs/common": "^10.4.16",
     "@next/bundle-analyzer": "^13.2.4",
     "@next/eslint-plugin-next": "^13.4.13",
